chore: refactor init to organize files slightly better
